# src/swagger/swagger.json
{
	"basePath": "/service",
	"consumes": [
		"application/json"
	],
	"definitions": {
		"UserRole": {
			"enum": [
				"admin",
				"user"
			],
			"type": "string"
		},
		"UserAttributes": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"$ref": "#/definitions/UserRole"
				},
				"password": {
					"type": "string",
					"minLength": 1,
					"x-nullable": true
				},
				"_id": {
					"type": "object"
				},
				"createdAt": {
					"type": "string",
					"format": "date-time"
				},
				"updatedAt": {
					"type": "string",
					"format": "date-time"
				}
			},
			"required": [
				"email",
				"name",
				"role",
				"_id",
				"createdAt",
				"updatedAt"
			],
			"type": "object"
		},
		"PaginationResponseUserAttributes[]": {
			"properties": {
				"count": {
					"type": "number",
					"format": "double"
				},
				"items": {
					"type": "array",
					"items": {
						"$ref": "#/definitions/UserAttributes"
					}
				}
			},
			"required": [
				"count",
				"items"
			],
			"type": "object"
		},
		"UserCreateRequest": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"$ref": "#/definitions/UserRole"
				},
				"password": {
					"type": "string"
				}
			},
			"required": [
				"email",
				"name",
				"role",
				"password"
			],
			"type": "object"
		},
		"UserRequest": {
			"properties": {
				"email": {
					"type": "string",
					"format": "email",
					"pattern": "^[a-zA-Z0-9_.+-]+\\x40[a-zA-Z0-9-]+\\.[a-zA-Z0-9-.]+$"
				},
				"name": {
					"type": "string",
					"minLength": 1
				},
				"role": {
					"$ref": "#/definitions/UserRole"
				},
				"password": {
					"type": "string",
					"minLength": 1,
					"x-nullable": true
				}
			},
			"required": [
				"email",
				"name",
				"role"
			],
			"type": "object"
		}
	},
	"info": {
		"title": "tsoa-api",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		}
	},
	"paths": {
		"/users/{id}": {
			"get": {
				"operationId": "Get",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string",
						"pattern": "^[A-Fa-f\\d]{24}$"
					}
				]
			},
			"put": {
				"operationId": "Update",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"400": {
						"description": "Bad request"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string",
						"pattern": "^[A-Fa-f\\d]{24}$"
					},
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserRequest"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"produces": [
					"application/json"
				],
				"responses": {
					"204": {
						"description": "No content"
					},
					"404": {
						"description": "Not Found"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"type": "string",
						"pattern": "^[A-Fa-f\\d]{24}$"
					}
				]
			}
		},
		"/users": {
			"get": {
				"operationId": "Search",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/PaginationResponseUserAttributes[]"
						}
					},
					"400": {
						"description": "Bad request"
					}
				},
				"tags": [
					"users"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 1
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"format": "int32",
						"type": "integer",
						"minimum": 1,
						"maximum": 100
					},
					{
						"in": "query",
						"name": "query",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"type": "string"
					},
					{
						"in": "query",
						"name": "sortDirection",
						"required": false,
						"type": "string",
						"enum": [
							"ASC",
							"DESC"
						],
						"pattern": "^ASC|DESC$"
					}
				]
			},
			"post": {
				"operationId": "Create",
				"produces": [
					"application/json"
				],
				"responses": {
					"200": {
						"description": "Ok",
						"schema": {
							"$ref": "#/definitions/UserAttributes"
						}
					},
					"400": {
						"description": "Bad request"
					}
				},
				"tags": [
					"users"
				],
				"security": [
					{
						"admin": []
					}
				],
				"parameters": [
					{
						"in": "body",
						"name": "body",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UserCreateRequest"
						}
					}
				]
			}
		}
	},
	"produces": [
		"application/json"
	],
	"swagger": "2.0",
	"securityDefinitions": {
		"jwt": {
			"type": "apiKey",
			"name": "access_token",
			"in": "header"
		}
	}
}